requirements.txt
requests==2.32.5
psycopg[binary]>=3.2.0
python-dotenv==1.2.1